+++ CHANGELOG.md
@@ -12,6 +12,7 @@ See [0Ver](https://0ver.org/).
 
 - Makes `_Nothing` a singleton
 - Fixes typos in documentation
+- Fixes that `assert_trace` was not catching containers from `@safe`-wrapped functions
 
 
 ## 0.15.0 aka The initial HKT release